Server/src/main/java/edu/sdccd/cisc191/template/MultiThreadedServer.java
package edu.sdccd.cisc191.template.server;

import java.io.IOException;
import java.net.ServerSocket;
import java.net.Socket;
import java.util.concurrent.ExecutorService;
import java.util.concurrent.Executors;

public class MultiThreadedServer {
    private ServerSocket serverSocket;
    private ExecutorService executorService;

    public MultiThreadedServer(int port, int poolSize) throws IOException {
        serverSocket = new ServerSocket(port);
        executorService = Executors.newFixedThreadPool(poolSize);
    }

    public void start() {
        while (true) {
            try {
                Socket clientSocket = serverSocket.accept();
                executorService.execute(new ClientHandler(clientSocket));
            } catch (IOException e) {
                e.printStackTrace();
            }
        }
    }
}
